# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me:
    get:
      operationId: getCurrentSession
      tags: [session]
      summary: Get the current authenticated session
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentSession"
        "401":
          description: Unauthorized
  /me/preferences:
    patch:
      operationId: updateUserPreferences
      tags: [session]
      summary: Update the current user's workspace preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserPreferencesInput"
      responses:
        "200":
          description: Updated user preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPreferences"
        "400":
          description: Invalid preferences
        "401":
          description: Unauthorized
  /organizations:
    post:
      operationId: createOrganization
      tags: [organizations]
      summary: Create an organization for the current user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationInput"
      responses:
        "201":
          description: Created organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "400":
          description: Invalid organization data
        "401":
          description: Unauthorized
  /organizations/{organizationId}:
    get:
      operationId: getOrganization
      tags: [organizations]
      summary: Get an organization the current user belongs to
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
      responses:
        "200":
          description: Organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "403":
          description: Organization access denied
        "404":
          description: Organization not found
    patch:
      operationId: updateOrganization
      tags: [organizations]
      summary: Update an organization
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationUpdate"
      responses:
        "200":
          description: Updated organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "400":
          description: Invalid organization data
        "403":
          description: Organization access denied
        "404":
          description: Organization not found
  /organizations/{organizationId}/dashboard:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get an empty-safe dashboard summary
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
        "403":
          description: Organization access denied
        "404":
          description: Organization not found
  /organizations/{organizationId}/audit-logs:
    get:
      operationId: listAuditLogs
      tags: [organizations]
      summary: List recent organization audit logs
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
      responses:
        "200":
          description: Audit logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLog"
        "403":
          description: Organization access denied
...
  /organizations/{organizationId}/modules:
    get:
      operationId: listOrganizationModules
      tags: [modules]
      summary: List module entitlements for an organization
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
      responses:
        "200":
          description: Organization module entitlements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrganizationModule"
        "403":
          description: Organization access denied
...
components:
  parameters:
...
    OrganizationId:
      name: organizationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    User:
      type: object
      required: [id, email, displayName]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        displayName:
          type: string
    Organization:
      type: object
      required:
        - id
        - legalNameEnglish
        - country
        - currency
        - timezone
        - defaultLanguage
        - vatRegistered
      properties:
        id:
          type: string
          format: uuid
        legalNameEnglish:
          type: string
        legalNameArabic:
          type: ["string", "null"]
        tradingNameEnglish:
          type: ["string", "null"]
        tradingNameArabic:
          type: ["string", "null"]
        businessType:
          type: ["string", "null"]
          enum:
            - establishment
            - limited_liability_company
            - joint_stock_company
            - professional_company
            - non_profit
            - other
            - null
        vatNumber:
          type: ["string", "null"]
        commercialRegistrationNumber:
          type: ["string", "null"]
        country:
          type: string
        city:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        streetName:
          type: ["string", "null"]
        postalCode:
          type: ["string", "null"]
        additionalNumber:
          type: ["string", "null"]
        buildingNumber:
          type: ["string", "null"]
        district:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        currency:
          type: string
        fiscalYearStart:
          type: string
        defaultLanguage:
          type: string
          enum: [en, ar]
        timezone:
          type: string
        vatRegistered:
          type: boolean
        numberFormat:
          type: string
          enum: [western, arabic]
        invoiceLanguage:
          type: string
          enum: [en, ar, bilingual]
        onboardingCompleted:
          type: boolean
        createdAt:
          type: string
          format: date-time
    OrganizationInput:
      type: object
      required: [legalNameEnglish]
      properties:
        legalNameEnglish:
          type: string
          minLength: 1
        legalNameArabic:
          type: ["string", "null"]
        tradingNameEnglish:
          type: ["string", "null"]
        tradingNameArabic:
          type: ["string", "null"]
        businessType:
          type: ["string", "null"]
          enum:
            - establishment
            - limited_liability_company
            - joint_stock_company
            - professional_company
            - non_profit
            - other
            - null
        vatNumber:
          type: ["string", "null"]
        commercialRegistrationNumber:
          type: ["string", "null"]
        country:
          type: string
          default: Saudi Arabia
        city:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        streetName:
          type: ["string", "null"]
        postalCode:
          type: ["string", "null"]
        additionalNumber:
          type: ["string", "null"]
        buildingNumber:
          type: ["string", "null"]
        district:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        currency:
          type: string
          default: SAR
        fiscalYearStart:
          type: string
          default: "01-01"
        defaultLanguage:
          type: string
          enum: [en, ar]
          default: en
        timezone:
          type: string
          default: Asia/Riyadh
        vatRegistered:
          type: boolean
          default: false
        numberFormat:
          type: string
          enum: [western, arabic]
          default: western
        invoiceLanguage:
          type: string
          enum: [en, ar, bilingual]
          default: bilingual
        onboardingCompleted:
          type: boolean
          default: false
    OrganizationUpdate:
      allOf:
        - $ref: "#/components/schemas/OrganizationInput"
    OrganizationMembership:
      type: object
      required: [organization, role]
      properties:
        organization:
          $ref: "#/components/schemas/Organization"
        role:
          type: string
          enum: [owner, admin, accountant, sales, purchasing, viewer]
    CurrentSession:
      type: object
      required: [user, organizations, preferences]
      properties:
        user:
          $ref: "#/components/schemas/User"
        organizations:
          type: array
          items:
            $ref: "#/components/schemas/OrganizationMembership"
        preferences:
          $ref: "#/components/schemas/UserPreferences"
    UserPreferences:
      type: object
      required: [language, appearance, density, sidebarCollapsed]
      properties:
        language:
          type: string
          enum: [en, ar]
        appearance:
          type: string
          enum: [light, dark, system]
        density:
          type: string
          enum: [compact, comfortable]
        sidebarCollapsed:
          type: boolean
        currentOrganizationId:
          type: ["string", "null"]
          format: uuid
    UserPreferencesInput:
      type: object
      properties:
        language:
          type: string
          enum: [en, ar]
        appearance:
          type: string
          enum: [light, dark, system]
        density:
          type: string
          enum: [compact, comfortable]
        sidebarCollapsed:
          type: boolean
        currentOrganizationId:
          type: ["string", "null"]
          format: uuid
    DashboardSummary:
      type: object
      required: [currency, revenue, receivables, expenses, netProfit, hasComparativeData, recentTransactions, customerCount, supplierCount, customerChecklist, itemCount, itemChecklist]
      properties:
        currency:
          type: string
        revenue:
          type: string
        receivables:
          type: string
        expenses:
          type: string
        netProfit:
          type: string
        hasComparativeData:
          type: boolean
        recentTransactions:
          type: array
          items:
            type: object
            required: [id, label, amount]
            properties:
              id:
                type: string
              label:
                type: string
              amount:
                type: string
        customerCount: { type: integer }
        supplierCount: { type: integer }
        customerChecklist:
          type: object
          required: [hasCustomers, complete]
          properties:
            hasCustomers: { type: boolean }
            complete: { type: boolean }
        itemCount: { type: integer }
        itemChecklist:
          type: object
          required: [hasItems, complete]
          properties:
            hasItems: { type: boolean }
            complete: { type: boolean }
...
    AuditLog:
      type: object
      required: [id, action, entityType, createdAt]
      properties:
        id:
          type: string
          format: uuid
        action:
          type: string
        entityType:
          type: string
        entityId:
          type: ["string", "null"]
        actorName:
          type: ["string", "null"]
        previousValues:
          type: ["object", "null"]
          additionalProperties: true
        newValues:
          type: ["object", "null"]
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
    ModuleDefinition:
      type: object
      required:
        - key
        - name
        - nameAr
        - description
        - descriptionAr
        - category
        - icon
        - route
        - status
        - sortOrder
        - dependencies
        - availableForActivation
        - features
      properties:
        key:
          type: string
          enum: [finance, fleet, projects, assets, intelligence, automate]
        name:
          type: string
        nameAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        category:
          type: string
          enum: [business, intelligence, automation]
        icon:
          type: string
        route:
          type: string
        status:
          type: string
          enum: [AVAILABLE, COMING_SOON, BETA, DISABLED]
        sortOrder:
          type: integer
        dependencies:
          type: array
          items:
            type: string
        availableForActivation:
          type: boolean
        features:
          type: array
          items:
            type: string
    OrganizationModule:
      type: object
      required: [module, enabled]
      properties:
        module:
          $ref: "#/components/schemas/ModuleDefinition"
        enabled:
          type: boolean
        activatedAt:
          type: ["string", "null"]
          format: date-time
...