# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryCustomerChecklist } from './dashboardSummaryCustomerChecklist';
import type { DashboardSummaryItemChecklist } from './dashboardSummaryItemChecklist';
import type { DashboardSummaryRecentTransactionsItem } from './dashboardSummaryRecentTransactionsItem';

export interface DashboardSummary {
  currency: string;
  revenue: string;
  receivables: string;
  expenses: string;
  netProfit: string;
  hasComparativeData: boolean;
  recentTransactions: DashboardSummaryRecentTransactionsItem[];
  customerCount: number;
  supplierCount: number;
  customerChecklist: DashboardSummaryCustomerChecklist;
  itemCount: number;
  itemChecklist: DashboardSummaryItemChecklist;
}